peer: the 'noplugin' build tag disables plugins
Go plugins are not supported on all platforms and architectures and
there are complications with different libc implementations. This tag
can be used to disable plugins on platforms supported by Go to avoid
these complications.
