+++ core/handlers/library/registry.go
@@ -8,8 +8,6 @@ package library
 
 import (
 	"fmt"
-	"os"
-	"plugin"
 	"reflect"
 	"sync"
 
@@ -126,113 +124,6 @@ func (r *registry) loadCompiled(handlerFactory string, handlerType HandlerType, 
 		}
 		r.validators[extraArgs[0]] = inst.(validation.PluginFactory)
 	}
-}
-
-// loadPlugin loads a pluggable handler
-func (r *registry) loadPlugin(pluginPath string, handlerType HandlerType, extraArgs ...string) {
-	if _, err := os.Stat(pluginPath); err != nil {
-		logger.Panicf(fmt.Sprintf("Could not find plugin at path %s: %s", pluginPath, err))
-	}
-	p, err := plugin.Open(pluginPath)
-	if err != nil {
-		logger.Panicf(fmt.Sprintf("Error opening plugin at path %s: %s", pluginPath, err))
-	}
-
-	if handlerType == Auth {
-		r.initAuthPlugin(p)
-	} else if handlerType == Decoration {
-		r.initDecoratorPlugin(p)
-	} else if handlerType == Endorsement {
-		r.initEndorsementPlugin(p, extraArgs...)
-	} else if handlerType == Validation {
-		r.initValidationPlugin(p, extraArgs...)
-	}
-}
-
-// initAuthPlugin constructs an auth filter from the given plugin
-func (r *registry) initAuthPlugin(p *plugin.Plugin) {
-	constructorSymbol, err := p.Lookup(authPluginFactory)
-	if err != nil {
-		panicWithLookupError(authPluginFactory, err)
-	}
-	constructor, ok := constructorSymbol.(func() auth.Filter)
-	if !ok {
-		panicWithDefinitionError(authPluginFactory)
-	}
-
-	filter := constructor()
-	if filter != nil {
-		r.filters = append(r.filters, filter)
-	}
-}
-
-// initDecoratorPlugin constructs a decorator from the given plugin
-func (r *registry) initDecoratorPlugin(p *plugin.Plugin) {
-	constructorSymbol, err := p.Lookup(decoratorPluginFactory)
-	if err != nil {
-		panicWithLookupError(decoratorPluginFactory, err)
-	}
-	constructor, ok := constructorSymbol.(func() decoration.Decorator)
-	if !ok {
-		panicWithDefinitionError(decoratorPluginFactory)
-	}
-	decorator := constructor()
-	if decorator != nil {
-		r.decorators = append(r.decorators, constructor())
-	}
-}
-
-func (r *registry) initEndorsementPlugin(p *plugin.Plugin, extraArgs ...string) {
-	if len(extraArgs) != 1 {
-		logger.Panicf("expected 1 argument in extraArgs")
-	}
-	factorySymbol, err := p.Lookup(pluginFactory)
-	if err != nil {
-		panicWithLookupError(pluginFactory, err)
-	}
-
-	constructor, ok := factorySymbol.(func() endorsement2.PluginFactory)
-	if !ok {
-		panicWithDefinitionError(pluginFactory)
-	}
-	factory := constructor()
-	if factory == nil {
-		logger.Panicf("factory instance returned nil")
-	}
-	r.endorsers[extraArgs[0]] = factory
-}
-
-func (r *registry) initValidationPlugin(p *plugin.Plugin, extraArgs ...string) {
-	if len(extraArgs) != 1 {
-		logger.Panicf("expected 1 argument in extraArgs")
-	}
-	factorySymbol, err := p.Lookup(pluginFactory)
-	if err != nil {
-		panicWithLookupError(pluginFactory, err)
-	}
-
-	constructor, ok := factorySymbol.(func() validation.PluginFactory)
-	if !ok {
-		panicWithDefinitionError(pluginFactory)
-	}
-	factory := constructor()
-	if factory == nil {
-		logger.Panicf("factory instance returned nil")
-	}
-	r.validators[extraArgs[0]] = factory
-}
-
-// panicWithLookupError panics when a handler constructor lookup fails
-func panicWithLookupError(factory string, err error) {
-	logger.Panicf(fmt.Sprintf("Plugin must contain constructor with name %s. Error from lookup: %s",
-		factory, err))
-}
-
-// panicWithDefinitionError panics when a handler constructor does not match
-// the expected function definition
-func panicWithDefinitionError(factory string) {
-	logger.Panicf(fmt.Sprintf("Constructor method %s does not match expected definition",
-		factory))
 }
 
 // Lookup returns a list of handlers with the given
